+++ src/features/chat/chat-services/chat-thread-service.ts
@@ -18,6 +18,7 @@ import {
 import { FindAllChatDocuments } from "./chat-document-service"
 import { deleteDocuments } from "@/features/chat/chat-services/azure-cog-search/azure-cog-vector-store"
 import { handleCosmosError } from "@/features/common/cosmos-error.ts"
+import { useChatContext } from "../chat-ui/chat-context"
 
 function threeMonthsAgo(): string {
   const date = new Date()
@@ -159,6 +160,24 @@ export const EnsureChatThreadIsForCurrentUser = async (chatThreadID: string) => 
   }
 
   return modelToSave[0]
+}
+
+export const UpsertSelectedPromptButton = async (selectedPromptButton: string) => {
+  const { id } = useChatContext()
+  try {
+    const threads: ChatThreadModel[] = await FindChatThreadByID(id)
+    threads.forEach(async thread => {
+      const itemToUpdate = {
+        ...thread,
+      }
+      itemToUpdate.selectedPrompt = selectedPromptButton
+      await UpsertChatThread({
+        ...itemToUpdate,
+      })
+    })
+  } catch (error) {
+    console.log("Prompt button not selected", error)
+  }
 }
 
 export const UpsertChatThread = async (chatThread: ChatThreadModel) => {
