+++ src/features/chat/chat-ui/chat-empty-state/chat-message-empty-state.tsx
@@ -7,52 +7,36 @@ import { ChatTypeSelector } from "./chat-type-selector"
 import { PromptButton } from "./prompt-buttons-UI"
 import { Card } from "@/features/ui/card"
 import { EasterEgg } from "./chat-easter-egg"
-import { FindChatThreadByID, UpsertChatThread } from "../../chat-services/chat-thread-service"
-import { ChatThreadModel } from "../../chat-services/models"
+import { UpsertSelectedPromptButton } from "../../chat-services/chat-thread-service"
 
 interface Prop {}
 
 export const ChatMessageEmptyState: FC<Prop> = () => {
-  const { setInput, id } = useChatContext()
+  const { setInput } = useChatContext()
   const [selectedPrompt, setSelectedPrompt] = useState<string | undefined>(undefined)
 
   const handlePromptSelected = async (prompt: string): Promise<void> => {
     setSelectedPrompt(prompt)
-
-    try {
-      setInput(prompt)
-      const threads: ChatThreadModel[] = await FindChatThreadByID(id)
-      threads.forEach(async thread => {
-        const itemToUpdate = {
-          ...thread,
-        }
-        itemToUpdate.selectedPrompt = prompt
-        await UpsertChatThread({
-          ...itemToUpdate,
-        })
-      })
-    } catch (error) {
-      console.log("Prompt button not selected", error)
-    }
+    void UpsertSelectedPromptButton(prompt)
+    setInput(prompt)
   }
-
   const { fileState }: { fileState: { showFileUpload: string } } = useChatContext()
   const { showFileUpload } = fileState
 
   return (
-    <div className="grid grid-cols-5 w-full items-center container overflow-auto mx-auto max-w-3xl justify-center max:h-5/6 p-4 gap-9 pb-[80px]">
+    <div className="max:h-5/6 container mx-auto grid w-full max-w-3xl grid-cols-5 items-center justify-center gap-9 overflow-auto p-4 pb-[80px]">
       <Card className="col-span-5 flex flex-col gap-2 p-5 ">
         <EasterEgg />
         <div className="flex flex-col gap-1">
-          <p className="text-sm text-text">Set the Sensitivity of your chat</p>
+          <p className="text-text text-sm">Set the Sensitivity of your chat</p>
           <ChatSensitivitySelector disable={false} />
         </div>
         <div className="flex flex-col gap-1">
-          <p className="text-sm text-text">Choose a conversation style</p>
+          <p className="text-text text-sm">Choose a conversation style</p>
           <ChatStyleSelector disable={false} />
         </div>
         <div className="flex flex-col gap-1">
-          <p className="text-sm text-text">How would you like to chat?</p>
+          <p className="text-text text-sm">How would you like to chat?</p>
           <ChatTypeSelector disable={false} />
         </div>
         {showFileUpload === "data" || showFileUpload === "audio" ? (
@@ -62,7 +46,7 @@ export const ChatMessageEmptyState: FC<Prop> = () => {
         )}
         <div className="flex flex-col gap-1">
           <br />
-          <p className="text-sm text-text">Try a suggested starter prompt...</p>
+          <p className="text-text text-sm">Try a suggested starter prompt...</p>
           <PromptButton onPromptSelected={handlePromptSelected} selectedPrompt={selectedPrompt} />
         </div>
       </Card>
